feat: agregar funcion platos a la api
la funcion platos recolecta la informacion de la base de datos de prueba

CARD ID: 66577e3b994e8802d61d0d63
